Use SessionDuration attribute to support SSO sessions of any length (#6)
Previously we were hardcoding the timespan of each temporary session token to
1 hour. But JumpCloud's SAML assertion usually contains a `SessionDuration`
attribute, which can be any length of time up to 12 hours. This change inspects
the XML for the `SessionDuration` and uses that to support whatever duration is
in the JumpCloud SSO configuration.

For Guild, the upshot is that security tokens will only need to be refreshed
after 12 hours, rather than after 1 hour.