Update copyright/license statement: project now licensed under MIT license (#5)
Per discussion last month, Donnie agreed that this project could be open-sourced after we had 2FA compliance in JumpCloud. As of late last week, everyone with AWS access also has 2FA enabled, and that will be enforced by IT on an ongoing basis. I think that means it's safe to release this as open source.