Edits to 1Password description
Very fired up about the new 1Password integration!!!

I've tried to make the description a little clearer.  I removed "guild-education" and "guild-dev", and added a subdomain name and profile name consistent with your earlier examples.

Any chance that y'all might tweak this implementation so that the item name isn't *required* to be "jumpcloud"?  :)

Thanks for the hard work!